requirements.txt
coverage==7.4.0
Flask==3.0.1
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
pg8000==1.30.4
SQLAlchemy==2.0.25
requests==2.31.0
marshmallow==3.20.2
Faker==22.5.1
python-dateutil~=2.8.2